refactor: separate confirmation-dialog from useDialog - reusable confirmation-dialog component - useDialog hook and context provider
